Add pagination to Employee module and fix type mappings
- Added PaginatedList helper class for server-side pagination
- Updated EmployeeController with pagination support (page, pageSize params)
- Updated Index view with pagination controls and page size selector
- Fixed decimal/double type mappings in ApplicationDbContext for database compatibility
- Added explicit column type mappings for float, decimal, and int columns

+++ src/HRMBT.Web/Models/Employee.cs
@@ -83,12 +83,12 @@ public class Employee
     public double? CarryForwardLeaves1 { get; set; }
 
     [Display(Name = "Year 2023 New")]
-    public double? Year2023New { get; set; }
+    public decimal? Year2023New { get; set; }
 
     [Display(Name = "Basic Salary")]
     [Range(0, double.MaxValue, ErrorMessage = "Basic salary must be greater than or equal to zero")]
     [DataType(DataType.Currency)]
-    public double? BasicSalary { get; set; }
+    public decimal? BasicSalary { get; set; }
 
     [Display(Name = "Apply Tax")]
     [StringLength(10)]
